+++ server/src/routes/faps/edit.routes.ts
@@ -48,29 +48,31 @@ export default (app: Router) => {
         '/rate/add',
         verifyUserToken,
         async (req, res) => {
-            try {
-                checkUserRoleOrErr(req, res, Roles.EXPERT)
-                const entityManager = AppDataSource.createEntityManager()
-
-                let date = new Date()
-                // @ts-ignore
-                date = date.getFullYear() + '-' + (date.getMonth() < 10 ? '0' + date.getMonth() : date.getMonth()) + '-' + date.getDate()
-
-                const query = "INSERT INTO `staff` (`id`, `medical_center_id`, `date`, `position`, `rate_full`, `rate_occupied`) VALUES (?, ?, ?, ?, ?, ?)"
-
+            const granted = checkUserRoleOrErr(req, res, Roles.EXPERT)
+            if (granted) {
                 try {
-                    const result = await entityManager.query(query,
-                        [null, req.body.medical_center_id, date, req.body.position, parseFloat(req.body.rate_full), parseFloat(req.body.rate_occupied)])
-                    res.json(result)
-                } catch (err) {
-                    await updateRatesInDataBase(req.body.medical_center_id)
-                    res.json({
-                        success: true
-                    })
+                    const entityManager = AppDataSource.createEntityManager()
+
+                    let date = new Date()
+                    // @ts-ignore
+                    date = date.getFullYear() + '-' + (date.getMonth() < 10 ? '0' + date.getMonth() : date.getMonth()) + '-' + date.getDate()
+
+                    const query = "INSERT INTO `staff` (`id`, `medical_center_id`, `date`, `position`, `rate_full`, `rate_occupied`) VALUES (?, ?, ?, ?, ?, ?)"
+
+                    try {
+                        const result = await entityManager.query(query,
+                            [null, req.body.medical_center_id, date, req.body.position, parseFloat(req.body.rate_full), parseFloat(req.body.rate_occupied)])
+                        res.json(result)
+                    } catch (err) {
+                        await updateRatesInDataBase(req.body.medical_center_id)
+                        res.json({
+                            success: true
+                        })
+                    }
+                } catch (e) {
+                    console.log(e)
+                    res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
                 }
-            } catch (e) {
-                console.log(e)
-                res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
             }
         }
     )
@@ -80,26 +82,28 @@ export default (app: Router) => {
         '/rate/update',
         verifyUserToken,
         async (req, res) => {
-            checkUserRoleOrErr(req, res, Roles.EXPERT)
-            try {
-                const entityManager = AppDataSource.createEntityManager()
+            const granted = checkUserRoleOrErr(req, res, Roles.EXPERT)
+            if (granted) {
+                try {
+                    const entityManager = AppDataSource.createEntityManager()
 
-                let date = new Date()
-                // @ts-ignore
-                date = date.getFullYear() + '-' + (date.getMonth() < 10 ? '0' + date.getMonth() : date.getMonth()) + '-' + date.getDate()
+                    let date = new Date()
+                    // @ts-ignore
+                    date = date.getFullYear() + '-' + (date.getMonth() < 10 ? '0' + date.getMonth() : date.getMonth()) + '-' + date.getDate()
 
-                const query: string = "UPDATE `staff` SET `date` = ?, `position` = ?, `rate_full` = ?, `rate_occupied` = ? WHERE `staff`.`id` = ?"
+                    const query: string = "UPDATE `staff` SET `date` = ?, `position` = ?, `rate_full` = ?, `rate_occupied` = ? WHERE `staff`.`id` = ?"
 
-                await entityManager.query(query, [date, req.body.position, req.body.rate_full, req.body.rate_occupied, req.body.id])
+                    await entityManager.query(query, [date, req.body.position, req.body.rate_full, req.body.rate_occupied, req.body.id])
 
-                await updateRatesInDataBase(req.body.medical_center_id)
+                    await updateRatesInDataBase(req.body.medical_center_id)
 
-                res.json({
-                    success: true
-                })
-            } catch (e) {
-                console.log(e)
-                res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
+                    res.json({
+                        success: true
+                    })
+                } catch (e) {
+                    console.log(e)
+                    res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
+                }
             }
         }
     )
@@ -109,18 +113,20 @@ export default (app: Router) => {
         '/rate/delete',
         verifyUserToken,
         async (req, res) => {
-            checkUserRoleOrErr(req, res, Roles.EXPERT)
-            try {
-                const entityManager = AppDataSource.createEntityManager()
+            const granted = checkUserRoleOrErr(req, res, Roles.EXPERT)
+            if (granted) {
+                try {
+                    const entityManager = AppDataSource.createEntityManager()
 
-                const query = "DELETE FROM `staff` WHERE `staff`.`id` = ?"
+                    const query = "DELETE FROM `staff` WHERE `staff`.`id` = ?"
 
-                await entityManager.query(query, [req.body.id])
-                await updateRatesInDataBase(req.body.medical_center_id)
+                    await entityManager.query(query, [req.body.id])
+                    await updateRatesInDataBase(req.body.medical_center_id)
 
-            } catch (e) {
-                console.log(e)
-                res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
+                } catch (e) {
+                    console.log(e)
+                    res.status(500).json({message: 'Что-то пошло не так, попробуйте снова'})
+                }
             }
         }
     )
